Move pool module to threading module
This change moves the pool module to the threading module, to make it
more widely available.
